move DeveloperOpenGraphTagsComponent to Developers namespace
