--- app/components/developer_open_graph_tags_component.rb
@@ -1,25 +1,0 @@
-class DeveloperOpenGraphTagsComponent < OpenGraphTagsComponent
-  def initialize(developer:)
-    @developer = developer
-  end
-
-  def title
-    [@developer.hero, "railsdevs"].join(" · ")
-  end
-
-  def description
-    @developer.bio
-  end
-
-  def url
-    developer_url(@developer)
-  end
-
-  def image
-    rails_blob_url(@developer.avatar) if @developer.avatar.attached?
-  end
-
-  def twitter
-    "@#{@developer.twitter}" if @developer.twitter.present?
-  end
-end
