+++ app/views/developers/show.html.erb
@@ -1,4 +1,4 @@
-<%= open_graph_tags component: DeveloperOpenGraphTagsComponent, developer: @developer %>
+<%= open_graph_tags component: Developers::OpenGraphTagsComponent, developer: @developer %>
 
 <div class="max-w-5xl mx-auto mt-0 lg:mt-16 pb-8 mb-16 bg-white shadow lg:rounded-lg overflow-hidden">
   <%= render CoverImageComponent.new(developer: @developer, classes: "lg:rounded-t-lg") %>
